OrderItem Deleting a reservation by name and phone number

+++ bookstore-cafe/src/main/java/springboot/bookstorecafe/services/ReservationService.java
@@ -89,17 +89,14 @@ public class ReservationService {
 			throw new RuntimeException("The reservation is no longer active.");
 		}
 
-//		if (reservation.getPerson().getLastName() != lastName
-//				|| reservation.getPerson().getPhoneNumber() != phoneNumber) {
-//
-//			throw new RuntimeException("Wrong last name or phone number! ");
-//		}
+		if (!reservation.getPerson().getLastName().equals(lastName)
+				|| !reservation.getPerson().getPhoneNumber().equals(phoneNumber)) {
+
+			throw new RuntimeException("Wrong last name or phone number! ");
+		}
 
 		reservation.getBookTable().setReservation(false);
 		reservationRepo.delete(reservation);
 	}
 
-//	public boolean isPersonDataMatch(String lastName, String phoneNumber) {
-//		return 
-//	}
 }
